#1 fix: seq_num 타입을 uint8_t에서 uint16_t로 변경
L3_MSG_MAX_SEQNUM(1024)과 실제 변수 타입(uint8_t, 최대 255) 불일치 수정
buildTxn, buildCnf의 seq 파라미터 타입도 동일하게 변경

+++ trader/L3_FSMmain.cpp
@@ -21,7 +21,7 @@ static uint8_t  coordId;
 static uint8_t  isSeller;
 static uint8_t  goods;
 static uint16_t price;
-static uint8_t  seq_num = 0;
+static uint16_t seq_num = 0;
 
 // coordinator로부터 수신한 값 
 static uint16_t rcvd_avg_price     = 0;
